Level loads contiguous array for bodies and parallax sprites Bodies are all inside a single union saved in one array. Parallax sprites are naturally uniform size (for the moment) although they are loaded on a per object basis cos the construction is so simple

+++ Graphics/Skins/StaticSkin.h
@@ -7,13 +7,13 @@
 class StaticSkin : public Skin
 {
     public:
-        StaticSkin(int _width, int _height);
+        StaticSkin(float _width, float _height);
         virtual ~StaticSkin();
         void vRender();
         SkinType getType(){return eStaticSkinType;}
     protected:
     private:
-        int width,height;
+        float width,height;
 };
 
 #endif // STATICSKIN_H
